test(httpserver): split production loopback coverage file
